+++ src/app/pathogen-notification/formly/configs/pathogen/submitting-facility.config.ts
@@ -29,7 +29,6 @@ import { TEXT_MAX_LENGTH } from '../../../legacy/common-utils';
 import { practitionerInfoFormConfigFields } from '../../../legacy/formly/configs/reusable/practitioner-info.config';
 import { contactsFormConfigFields } from '../../../legacy/formly/configs/reusable/contacts.config';
 import { ErrorDialogService } from '../../../services/error-dialog.service';
-import { isContactPointUseDisabled } from '../../../utils/pathogen-notification-mapper';
 
 export const submittingFacilityFields = (countryCodeDisplays: CodeDisplay[], dialogService: ErrorDialogService): FormlyFieldConfig[] => {
   let isResettingCheckbox = false;
@@ -128,7 +127,7 @@ export const submittingFacilityFields = (countryCodeDisplays: CodeDisplay[], dia
     },
     ...practitionerInfoFormConfigFields(),
     { template: '<h2>Kontaktmöglichkeiten</h2>' },
-    ...contactsFormConfigFields(true, isContactPointUseDisabled()),
+    ...contactsFormConfigFields(true),
   ];
 };