+++ src/app/pathogen-notification/legacy/formly-options-lists.ts
@@ -20,6 +20,7 @@ import ResultEnum = MethodPathogenDTO.ResultEnum;
 import ReportStatusEnum = NotificationLaboratoryCategory.ReportStatusEnum;
 import ResistanceResultEnum = ResistanceDTO.ResistanceResultEnum;
 import ResistanceGeneResultEnum = ResistanceGeneDTO.ResistanceGeneResultEnum;
+import NotificationIdReferenceEnum = NotificationLaboratoryCategory.NotificationIdReferenceEnum;
 
 export const RESIDENCE_ADDRESS_TYPE_OPTION_LIST = [
   { value: AddressType.Primary, label: 'Hauptwohnung' },
@@ -50,6 +51,12 @@ export const REPORT_STATUS_OPTION_LIST = [
   { value: ReportStatusEnum.Final, label: 'Endgültig' },
   { value: ReportStatusEnum.Preliminary, label: 'Vorläufig' },
   { value: ReportStatusEnum.Amended, label: 'Ergänzung oder Korrektur' },
+];
+
+export const NOTIFICATION_ID_REFERENCE_LIST = [
+  { value: NotificationIdReferenceEnum.NoReference, label: 'Kein Verweis' },
+  { value: NotificationIdReferenceEnum.RelatesToOwnFacility, label: 'Meldung meiner Einrichtung' },
+  { value: NotificationIdReferenceEnum.RelatesToOtherFacility, label: 'Meldung einer anderen Einrichtung' },
 ];
 
 export const RESISTANCE_RESULT_OPTION_LIST = [